Re-Implemented TaskAdapter using the RecyclerViewAdapter<T> class
This class was created by me to simplyfied the creation of RecyclerView adapters.

The onDeleteSuccess signature changed to pass just the deleted task and not the whole list.

This change improve the performance of the app (it's not a huge app but it's good practice anyway).

Besides, before the animation of deleted one task was worse becase we replaced the old list with the modified list and now we specify the item we want to remove.

There are also changes to the naming convention (delete -> remove).